+++ src/pages/advanced-search/index.tsx
@@ -1,9 +1,10 @@
 import React from "react";
-
 import AdvancedSearch from "~/src/components/AdvancedSearch/AdvancedSearch";
 import Layout from "~/src/components/Layout/Layout";
 import { languagesFetcher } from "../../lib/api/SearchApi";
 import { ApiLanguageResponse } from "../../types/LanguagesQuery";
+import Error from "../_error";
+
 export async function getServerSideProps() {
   // Fetch all languages
   const languageResponse: ApiLanguageResponse = await languagesFetcher();
@@ -15,7 +16,11 @@ export async function getServerSideProps() {
 }
 
 const AdvancedSearchPage: React.FC<any> = (props) => {
-  const { languages } = props;
+  const { languages, status } = props;
+
+  if (status !== 200) {
+    return <Error statusCode={status} />;
+  }
 
   return (
     <Layout>